+++ scanner/test/fixtures/ruby-php-structural/clean/users_controller.rb
@@ -1,0 +1,6 @@
+class UsersController < ApplicationController
+  def index
+    @users = User.where('name = ?', params[:name])
+    render json: @users
+  end
+end