+++ scanner/test/fixtures/ruby-php-structural/vulnerable/users_controller.rb
@@ -1,0 +1,7 @@
+class UsersController < ApplicationController
+  def index
+    name = params[:name]
+    @users = User.where("name = '#{name}'")
+    render json: @users
+  end
+end